Enable strict warnings-as-errors, hidden symbol visibility, and quality/security fixes
Build: add -Wall -Wextra -Wpedantic -Wshadow -Wconversion -Wsign-conversion
-Wcast-qual -Wpointer-arith -Wstrict-prototypes -Wmissing-prototypes -Wformat=2
-Wundef -Wvla -Werror to the Makefile and flake; compile with
-fvisibility=hidden and export only the doorman_* API via a pragma in the
header. Build Objective-C under ARC in both build systems.

Fixes surfaced/hardened:
- constant-time compare: avoid narrowing compound assignment
- PAM: read PAM_USER without a qualifier-discarding cast
- provisioning: reject caller fields that are not valid UTF-8 (would throw
  when placed in an argument array)
- session launch: cast gid for initgroups (sign-conversion)
- CLI: mask terminal ECHO with tcflag_t; parse uid/gid with strtoul + range
  checks instead of atoi; drop a const-discarding cast on --exec

+++ doorman/src/sessions.m
@@ -246,7 +246,7 @@ doorman_result_t doorman_open_session(doorman_handle_t *handle,
         /* --- child: POSIX-only from here to exec --- */
         if (privileged) {
             if (setgid(gid) != 0) _exit(127);
-            if (initgroups(name_dup, gid) != 0) _exit(127);
+            if (initgroups(name_dup, (int)gid) != 0) _exit(127);
             if (setuid(uid) != 0) _exit(127);
             /* Defence in depth: if we dropped to a non-root uid, regaining root
              * must be impossible; bail out if it somehow is not. */
